# contracts/api-contract.yaml
# doc_type: api-contract
# service: kart-inventory-service
# status: approved (+ one addendum, see "Implementation Addendum" below)
# generated_by: api-design-agent
# source: docs/services/kart-inventory-service/requirement-spec.md,
#         docs/services/kart-inventory-service/edge-cases.md,
#         docs/services/kart-inventory-service/design-decisions.md,
#         docs/services/kart-inventory-service/architecture.md,
#         docs/services/kart-inventory-service/ddd-model.md
#
# Vendored, read-only copy of the approved contract from
# kart-platform/docs/services/kart-inventory-service/api-contract.yaml - see contracts/README.md
# for the vendoring convention. Update it only by re-copying the upstream file, never by editing
# it directly here - EXCEPT for the one addendum below, which fills a real gap in the upstream
# artifact rather than drifting from it.
#
# Implementation Addendum (added here, not upstream): the upstream contract never defines a REST
# path for INV-8 (ReplenishStock) even though tickets.md/architecture.md both require one
# ("Admin -> Inventory's own replenishment write path (manual trigger), Sync REST, internal
# client-credentials call"). POST /inventory/replenish below is added to close that gap, following
# this contract's own conventions exactly (Problem error shape, versioning, auth style). Flagged
# here explicitly rather than silently invented - see this repo's README for the full note.
#
# Second Implementation Addendum (Inventory & Stock Management flow, added here, not upstream):
# four more real gaps found during that flow's audit, closed the same way (AdminOnly-gated,
# Problem error shape, versioning/auth style unchanged) - POST /inventory/provision (onboards a
# brand-new (warehouseId, sku) row - previously only reachable via a migration seed), PATCH
# /inventory/{warehouseId}/{sku}/threshold (the "Low Stock Threshold" stage - previously
# settable only at provision time), POST /inventory/{warehouseId}/{sku}/reconcile (the "Stock
# Audit/Reconciliation" and "Update Qty" stages - did not exist at all), and GET
# /inventory/low-stock (the "Reorder Alert" dashboard view - previously a LogWarning only, never
# queryable). The Reservation.status enum also gained `committed` (the "Deduct (Order Confirmed)"
# stage - a reservation opts out of the TTL sweep once its order is paid, but must remain
# releasable on a later cancellation).
#
# This contract has two surfaces: a public/gateway-reachable REST API
# (reserve, release, read, replenish) and one internal gRPC service consumed only by
# Cart's checkout-time lazy validation call (kart-cart-service/design-decisions.md,
# "Resilience Pattern for Checkout-Time Stock/Price Validation" - resolved on
# this service's side in architecture.md's "API Surface Consistency Note").

openapi: 3.0.3
info:
  title: kart-inventory-service API
  version: 1.0.0
  description: >
    Stock truth per warehouse/SKU, synchronous reservation holds with TTL,
    and oversell prevention via row-level pessimistic locking (BRD S5.2,
    S6.1). The platform's highest-contention service (PLATFORM_BLUEPRINT.md).

servers:
  - url: /v1

security:
  - clientCredentials: []

paths:
  /inventory/reserve:
    post:
      operationId: reserveStock
      summary: Reserve stock for an order line item (synchronous saga step 1).
      description: >
        Called synchronously by Order as the first Order Saga step (BRD
        S12.1; ADR-0009 confirms this is genuine synchronous RPC, not an
        async hop). Locks the target SKU's warehouse stock row(s)
        (SELECT ... FOR UPDATE), attempting single-warehouse-first
        satisfaction, falling back to a multi-warehouse split only if no
        single warehouse can fully satisfy the line (requirement-spec.md
        Decision 3; ddd-model.md's Reservation/WarehouseAllocation shape).
        On success, publishes InventoryReserved after durable commit; on
        insufficient stock, publishes InventoryReservationFailed and
        returns 409 - there is no backorder path (requirement-spec.md
        Decision 4). A bounded lock_timeout (design-decisions.md,
        "Resilience Budget") returns 503 rather than queuing indefinitely
        on a hot SKU.
      security:
        - clientCredentials: [service:order]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [orderId, sku, qty]
              properties:
                orderId: { type: string, format: uuid }
                sku: { type: string }
                qty: { type: integer, minimum: 1 }
      responses:
        '201':
          description: Reservation created; InventoryReserved published.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Reservation' }
        '409':
          description: >
            Insufficient stock across all candidate warehouses;
            InventoryReservationFailed published. No backorder/waitlist
            path exists (requirement-spec.md Decision 4).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
        '503':
          description: >
            Lock-wait timeout on a hot SKU (design-decisions.md, "Resilience
            Budget") - retryable, distinct from a genuine 409 oversell
            failure.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }

  /inventory/release:
    post:
      operationId: releaseReservation
      summary: Release a held reservation (explicit call or saga compensation).
      description: >
        Idempotent, keyed on reservationId (ddd-model.md's ReservationStatus
        terminal-state machine) - a release against an already-Released or
        already-Expired reservation is a no-op that still returns 200 and
        does not re-publish InventoryReleased (requirement-spec.md S4;
        edge-cases.md, "Compensating-transaction rollback correctness").
        Reachable via this endpoint, by consuming OrderCancelled, by
        consuming OrderCompensationTriggered, or by the internal TTL sweep -
        all four converge on this same state-checked write path.
      security:
        - clientCredentials: [service:order]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reservationId]
              properties:
                reservationId: { type: string, format: uuid }
      responses:
        '200':
          description: >
            Released (or already-terminal, idempotent no-op).
            InventoryReleased published on the transition; not republished
            if already terminal.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Reservation' }
        '404':
          description: reservationId not found.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }

  /inventory/{sku}:
    get:
      operationId: getStockLevel
      summary: Current available stock level for a SKU (public read path).
      description: >
        Informational display read, not a source of truth for any
        allocation decision - reservation/release/replenishment always take
        a fresh SELECT ... FOR UPDATE regardless of this endpoint's cache
        state (design-decisions.md, "Read-Path Caching Strategy"). Backed
        by a short-TTL Redis cache-aside; a cache miss falls back to
        PostgreSQL.
      security: []
      parameters:
        - name: sku
          in: path
          required: true
          schema: { type: string }
        - name: warehouseId
          in: query
          required: false
          schema: { type: string }
          description: Omit for platform-wide available quantity summed across warehouses.
      responses:
        '200':
          description: Stock level.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StockLevel' }
        '404':
          description: sku not found.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }

  /inventory/replenish:
    post:
      operationId: replenishStock
      summary: Replenish warehouse stock (manual admin path or automated threshold trigger).
      description: >
        IMPLEMENTATION ADDENDUM - see file header. Backs INV-8: credits an
        existing warehouse_stock row via the same SELECT ... FOR UPDATE
        write path reservation/release use (requirement-spec.md Decision 5;
        edge-cases.md "Replenishment racing an active reservation"), then
        publishes InventoryReplenished. The (warehouseId, sku) row must
        already exist (provisioned at initial load) - this endpoint credits
        stock, it does not onboard new warehouse/SKU master data (ddd-model.md
        Modeling Decision 2; tickets.md's flagged gap).
      security:
        - clientCredentials: [admin, service:inventory-replenishment-trigger]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [warehouseId, sku, qtyAdded]
              properties:
                warehouseId: { type: string }
                sku: { type: string }
                qtyAdded: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Replenished; InventoryReplenished published.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StockLevel' }
        '404':
          description: (warehouseId, sku) has never been provisioned.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }

  /inventory/provision:
    post:
      operationId: provisionWarehouseStock
      summary: Onboard a brand-new (warehouseId, sku) row (Inventory & Stock Management flow).
      description: >
        IMPLEMENTATION ADDENDUM - see file header. Previously only reachable via a migration
        seed (ddd-model.md Modeling Decision 2). AdminOnly - master-data onboarding, distinct
        from POST /inventory/replenish, which only credits an already-provisioned row.
      security:
        - clientCredentials: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [warehouseId, sku, initialQty, replenishmentThreshold, targetStockingLevel]
              properties:
                warehouseId: { type: string }
                sku: { type: string }
                initialQty: { type: integer, minimum: 0 }
                replenishmentThreshold: { type: integer, minimum: 0 }
                targetStockingLevel: { type: integer, minimum: 1 }
      responses:
        '201':
          description: Provisioned; WarehouseStockProvisioned published.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StockLevel' }
        '400':
          description: Already provisioned, or invalid quantities/thresholds.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }

  /inventory/{warehouseId}/{sku}/threshold:
    patch:
      operationId: updateReplenishmentThreshold
      summary: Update a warehouse/SKU's low-stock threshold and target stocking level (Inventory & Stock Management flow's "Low Stock Threshold" stage).
      description: >
        IMPLEMENTATION ADDENDUM - see file header. Config-only change (no InventoryOutboxEvent) -
        a subsequent write crossing the new threshold raises LowStockDetected as usual.
      security:
        - clientCredentials: [admin]
      parameters:
        - name: warehouseId
          in: path
          required: true
          schema: { type: string }
        - name: sku
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [replenishmentThreshold, targetStockingLevel]
              properties:
                replenishmentThreshold: { type: integer, minimum: 0 }
                targetStockingLevel: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Threshold updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StockLevel' }
        '404':
          description: (warehouseId, sku) has never been provisioned.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }

  /inventory/{warehouseId}/{sku}/reconcile:
    post:
      operationId: reconcileStock
      summary: Overwrite AvailableQty to an admin-submitted physical count (Inventory & Stock Management flow's "Stock Audit/Reconciliation" and "Update Qty" stages).
      description: >
        IMPLEMENTATION ADDENDUM - see file header. Unlike replenish/reserve, this write is
        authoritative in either direction (can decrease stock, e.g. shrinkage) - publishes
        InventoryReconciled with the signed variance (countedQty - previous AvailableQty).
      security:
        - clientCredentials: [admin]
      parameters:
        - name: warehouseId
          in: path
          required: true
          schema: { type: string }
        - name: sku
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [countedQty, reason]
              properties:
                countedQty: { type: integer, minimum: 0 }
                reason: { type: string }
      responses:
        '200':
          description: Reconciled; InventoryReconciled published.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StockReconciliationResult' }
        '404':
          description: (warehouseId, sku) has never been provisioned.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }

  /inventory/low-stock:
    get:
      operationId: getLowStock
      summary: List every SKU/warehouse currently below its own replenishment threshold (Inventory & Stock Management flow's "Reorder Alert" dashboard view).
      description: >
        IMPLEMENTATION ADDENDUM - see file header. Previously this signal was a LogWarning only,
        never queryable. Public read (CanRead unconditional), same posture as GET /inventory/{sku}.
      security: []
      parameters:
        - name: warehouseId
          in: query
          required: false
          schema: { type: string }
      responses:
        '200':
          description: Every currently-below-threshold (warehouseId, sku) row, empty array if none.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/StockLevel' }

  /inventory/orders/{orderId}/allocations:
    get:
      operationId: getOrderAllocations
      summary: Read-only warehouse-allocation view for an order (Order Management (Admin) flow's "Assign Warehouse" step).
      description: >
        Warehouse allocation itself stays fully automatic inside the reserve
        saga (requirement-spec.md Decision 3's single-warehouse-first/
        multi-warehouse-fallback logic) - this endpoint is admin visibility
        into a decision the system already made, not a new admin override.
        Called by kart-admin-service when rendering an order's detail screen.
        Returns every reservation for the order regardless of status
        (reserved/released/expired), so an admin can see which warehouse(s)
        fulfilled a line item even after the reservation has since ended.
      security:
        - clientCredentials: [admin]
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Every reservation (any status) for this order, empty array if none were ever made.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Reservation' }

components:
  securitySchemes:
    clientCredentials:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: https://kart-identity-service/connect/token
          scopes:
            service:order: Order Service's own client-credentials principal
            admin: Admin Service's operator principal
            service:inventory-replenishment-trigger: internal automated reorder trigger principal

  schemas:
    Reservation:
      type: object
      required: [reservationId, orderId, sku, qty, status, allocations]
      properties:
        reservationId: { type: string, format: uuid }
        orderId: { type: string, format: uuid }
        sku: { type: string }
        qty: { type: integer }
        status:
          type: string
          enum: [reserved, committed, released, expired]
          description: >
            committed (Inventory & Stock Management flow's "Deduct (Order Confirmed)" stage,
            added here) means the order backing this reservation has been paid - it is no longer
            subject to the TTL sweep, but remains releasable on a later admin cancellation.
        allocations:
          type: array
          description: One entry per warehouse this reservation drew from (ddd-model.md's WarehouseAllocation) - usually one, more than one only on the multi-warehouse fallback.
          items:
            type: object
            properties:
              warehouseId: { type: string }
              qty: { type: integer }
        expiresAt:
          type: string
          format: date-time
          description: Creation time + 15-minute TTL (requirement-spec.md Decision 2).

    StockLevel:
      type: object
      required: [sku, availableQty]
      properties:
        sku: { type: string }
        warehouseId:
          type: string
          nullable: true
          description: Present only when the request scoped to one warehouse.
        availableQty: { type: integer }

    StockReconciliationResult:
      type: object
      required: [sku, warehouseId, previousQty, countedQty, variance]
      properties:
        sku: { type: string }
        warehouseId: { type: string }
        previousQty: { type: integer }
        countedQty: { type: integer }
        variance:
          type: integer
          description: countedQty - previousQty. Positive = found more than the ledger showed, negative = shrinkage.

    Problem:
      type: object
      description: RFC 7807-style problem details.
      required: [code, message]
      properties:
        code:
          type: string
          example: insufficient_stock
        message:
          type: string
        details:
          type: object
          additionalProperties: true

# =============================================================================
# Internal gRPC surface - InventoryAvailabilityService
#
# Consumed only by kart-cart-service's checkout-time lazy validation call
# (kart-cart-service/design-decisions.md, "Resilience Pattern for
# Checkout-Time Stock/Price Validation" - timeout-guarded, circuit-breaker
# wrapped, fails open on Cart's side). Read-only: never reserves, never
# takes a write lock, reads through the same short-TTL cache-aside path as
# GET /inventory/{sku} above - this is a second protocol for the same read
# concern, not a second source of truth (architecture.md's "API Surface
# Consistency Note - Resolved").
#
# proto3 service definition (see src/Api/Protos/inventory_availability.proto
# for the actual compiled file used by this implementation):
#
#   service InventoryAvailabilityService {
#     rpc CheckAvailability(CheckAvailabilityRequest) returns (CheckAvailabilityResponse);
#   }
#
#   message CheckAvailabilityRequest {
#     string sku = 1;
#     int32 requested_qty = 2;
#   }
#
#   message CheckAvailabilityResponse {
#     string sku = 1;
#     int32 available_qty = 2;
#     bool sufficient = 3;  // available_qty >= requested_qty at read time -
#                           // advisory only, re-checked authoritatively by
#                           // POST /inventory/reserve's own lock at checkout
# }
# =============================================================================